refactor and fixes
switch to https://github.com/almottier/TapoP100 because of new firmware
use i3block interval=persist instead of config caching for tp-plugs
refactor $PATH

+++ scripts/i3blocks/datescript.sh
@@ -1,30 +1,35 @@
 #!/bin/bash
 LANG=bg_BG.UTF-8
-thestr=" `date '+ %a %d %b %Y %H:%M:%S'`"
-full="$thestr"
-short="$thestr"
-color="#FFFFFF"
+
+if [ "${BLOCK_INTERVAL:-}" == "persist" ]; then
+  while :; do
+    echo "$(date '+ %a %d %b %Y %H:%M:%S')"
+    sleep 1
+  done
+  exit
+fi
 
 case $BLOCK_BUTTON in
-  1)
-    if ! pgrep gsimplecal 1>/dev/null 2>/dev/null; then
-      nohup gsimplecal 1>/dev/null 2>/dev/null &
-    fi
+1)
+  if ! pgrep gsimplecal 1>/dev/null 2>/dev/null; then
+    nohup gsimplecal 1>/dev/null 2>/dev/null &
+  fi
   ;;
-  3)
-    killall gsimplecal 1>/dev/null 2>/dev/null
+3)
+  killall gsimplecal 1>/dev/null 2>/dev/null
   ;;
-  4)
-    if pgrep gsimplecal 1>/dev/null 2>/dev/null; then
-      gsimplecal next_month
-    fi
+4)
+  if pgrep gsimplecal 1>/dev/null 2>/dev/null; then
+    gsimplecal next_month
+  fi
   ;;
-  5)
-    if pgrep gsimplecal 1>/dev/null 2>/dev/null; then
-      gsimplecal prev_month
-    fi
+5)
+  if pgrep gsimplecal 1>/dev/null 2>/dev/null; then
+    gsimplecal prev_month
+  fi
   ;;
 esac
-echo $full
-echo $short
-echo $color
+thestr="$(date '+ %a %d %b %Y %H:%M:%S')"
+echo "${thestr}"
+echo "${thestr}"
+echo "#FFFFFF"
